requirements.txt
streamlit==0.78.0
pillow
blinker
torchvision==0.2.1 
http://download.pytorch.org/whl/cpu/torch-1.0.0-cp36-cp36m-linux_x86_64.whl